refactor: entity에서 domain으로 패키지 명 변경
