+++ src/main/java/roomescape/time/domain/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.time.entity;
+package roomescape.time.domain;
 
 public class ReservationTime {
