+++ src/test/java/roomescape/reservation/repository/JdbcReservationRepositoryTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.reservation.entity.Reservation;
-import roomescape.time.entity.ReservationTime;
+import roomescape.reservation.domain.Reservation;
+import roomescape.time.domain.ReservationTime;
 
 @JdbcTest
 class JdbcReservationRepositoryTest {
